A spec/support/ directory, will be reused by openshift repo
