+++ spec/spec_helper.rb
@@ -8,17 +8,6 @@ VCR.configure do |config|
   config.cassette_library_dir = File.join(ManageIQ::Providers::Kubernetes::Engine.root, 'spec/vcr_cassettes')
 end
 
-# Helps constructing inputs similar to kubeclient results
-module ArrayRecursiveOpenStruct
-  def array_recursive_ostruct(hash)
-    RecursiveOpenStruct.new(hash, :recurse_over_arrays => true)
-  end
-end
-
-RSpec.configure do |c|
-  c.include ArrayRecursiveOpenStruct
-  c.extend ArrayRecursiveOpenStruct
-end
-
 Dir[Rails.root.join("spec/shared/**/*.rb")].each { |f| require f }
+# This repo's spec/support/ is also used by manageiq-providers-openshift.
 Dir[ManageIQ::Providers::Kubernetes::Engine.root.join("spec/support/**/*.rb")].each { |f| require f }
