+++ spec/support/array_recursive_ostruct.rb
@@ -1,0 +1,11 @@
+# Helps constructing inputs similar to kubeclient results
+module ArrayRecursiveOpenStruct
+  def array_recursive_ostruct(hash)
+    RecursiveOpenStruct.new(hash, :recurse_over_arrays => true)
+  end
+end
+
+RSpec.configure do |c|
+  c.include ArrayRecursiveOpenStruct
+  c.extend ArrayRecursiveOpenStruct
+end
